move components folder to src
