+++ src/components/PinkButton.css
